refactor(cloud): share cloud domain registry via notebook-cloud-transport
Lift CloudRegistry/CredentialRef/ResolvedCloudDomain out of runt-mcp into a
registry feature on notebook-cloud-transport so the daemon's hosted-room
bridge resolves the same machine-local cloud-domains.toml as MCP hosted
sessions. Adds parse_hosted_url and a generic NTERACT_CLOUD_REGISTRY env
override (NTERACT_MCP_CLOUD_REGISTRY still honored).
