+++ crates/notebook-cloud-transport/Cargo.toml
@@ -9,15 +9,25 @@ license.workspace = true
 [lints]
 workspace = true
 
+[features]
+# Machine-local cloud domain registry (routing + credential references).
+registry = ["dep:serde", "dep:toml", "dep:url", "dep:dirs", "dep:uuid", "dep:runt-workspace"]
+
 [dependencies]
 anyhow = { workspace = true }
+dirs = { workspace = true, optional = true }
 futures-util = { workspace = true }
 notebook-protocol = { path = "../notebook-protocol" }
 notebook-wire = { path = "../notebook-wire" }
+runt-workspace = { path = "../runt-workspace", optional = true }
+serde = { workspace = true, optional = true }
 serde_json = { workspace = true }
 tokio = { workspace = true, features = ["full"] }
 tokio-tungstenite = { workspace = true, features = ["native-tls"] }
+toml = { workspace = true, optional = true }
 tracing = { workspace = true }
+url = { workspace = true, optional = true }
+uuid = { workspace = true, optional = true }
 
 [dev-dependencies]
 tokio = { workspace = true, features = ["full", "test-util"] }
