+++ crates/runt-mcp/Cargo.toml
@@ -18,7 +18,7 @@ repr-llm = { path = "../repr-llm" }
 runtimed-client = { path = "../runtimed-client" }
 runtimed-outputs = { path = "../runtimed-outputs" }
 notebook-protocol = { path = "../notebook-protocol" }
-notebook-cloud-transport = { path = "../notebook-cloud-transport" }
+notebook-cloud-transport = { path = "../notebook-cloud-transport", features = ["registry"] }
 notebook-sync = { path = "../notebook-sync" }
 notebook-doc = { path = "../notebook-doc" }
 runtime-doc = { path = "../runtime-doc" }
